refactor image info flag
